moved identity tasks under tasks/identity, modified hardhat config, updated tests
