+++ hardhat.config.js
@@ -1,12 +1,12 @@
-require("@typechain/hardhat");
-require("@nomiclabs/hardhat-ethers");
-require("@openzeppelin/hardhat-upgrades");
+require('@typechain/hardhat');
+require('@nomiclabs/hardhat-ethers');
+require('@openzeppelin/hardhat-upgrades');
 
 const privateKey = process.env.DEPLOYER_ACCOUNT || '0x011967d88c6b79116bb879d4c2bc2c3caa23569edd85dfe0bc596846837bbc8e';
 const host = process.env.BLOCKCHAIN_HOST || 'blockchain_node';
 const port = process.env.BLOCKCHAIN_PORT || 7545;
 
-const devaddress = 'http://' + host + ':' + port
+const devaddress = 'http://' + host + ':' + port;
 
 // You need to export an object to set up your config
 // Go to https://hardhat.org/config/ to learn more
@@ -14,23 +14,17 @@ const devaddress = 'http://' + host + ':' + port
 let defaultNetwork;
 if (process.env.MODE === 'e2e' || process.env.MODE === 'zappdev') {
     defaultNetwork = 'development';
-}else{
+} else {
     defaultNetwork = 'ynet';
 }
 
 const config = {
     solidity: {
         compilers: [
-            {
-                version: "0.8.0",
-            },
-            {
-                version: "0.8.4",
-            },
-            {
-                version: "0.8.7",
-            }
-        ],
+            {version: '0.8.0'},
+            {version: '0.8.4'},
+            {version: '0.8.7'}
+        ]
     },
     paths: {
         artifacts:'./hardhat/build',
@@ -42,13 +36,13 @@ const config = {
         development: {
             url: devaddress,
             accounts:
-            [privateKey],
+            [privateKey]
         },
         ynet: {
             url: 'http://ynet.point.space:44444',
             accounts:
-                ['ea2a5e73b526b8a5f60c7f19719b6abe71f054721a8a367fff0a9e2cb07e1080'],
-        },
+                ['ea2a5e73b526b8a5f60c7f19719b6abe71f054721a8a367fff0a9e2cb07e1080']
+        }
     },
     defaultNetwork: defaultNetwork
 };
