+++ hardhat/hardhat.config.ts
@@ -11,10 +11,10 @@ import "./tasks/importer/identity";
 import "./tasks/importer/blog";
 import "./tasks/importer/pointSocial";
 import "./tasks/importer/sms";
-import "./tasks/importer/identity-update";
-import './tasks/importer/identity-add-deployer.ts';
-import './tasks/importer/identity-remove-deployer.ts';
-import './tasks/importer/identity-list-deployers.ts';
+import "./tasks/identity/identity-update-contract";
+import './tasks/identity/identity-add-deployer.ts';
+import './tasks/identity/identity-remove-deployer.ts';
+import './tasks/identity/identity-list-deployers.ts';
 
 const ethers = require('ethers');
 const keystore = {"phrase":"observe valid excite index skill drink argue envelope domain second ten hybrid"};
@@ -71,9 +71,7 @@ const config: HardhatUserConfig = {
             url: devaddress,
             accounts:
             [
-                privateKey,
-                '011967d88c6b79116bb879d4c2bc2c3caa23569edd85dfe0bc596846837bbc8e',
-                '22a316b515a68d4851087571bd5ff051f5ec3f13b28997fb80a8de055052514e'
+                privateKey
             ]
         },
         ynet: {
